requirements.txt
streamlit>=1.28.0
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.0.0
plotly>=5.15.0
wordcloud>=1.9.0
matplotlib>=3.7.0